third_party/WebKit/LayoutTests/http/tests/security/suborigins/multiple-suborigins-disallowed.html

<!DOCTYPE html>
<html>
<head>
    <title>Multiple Suborigins Disallowed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<script>

window.onmessage = function() {
    var secret = '';
    try {
        secret = iframe.contentWindow.secret;
        assert_unreached();
    } catch(e) {
        assert_equals(secret, "", "The parent frame should never be able to get the secret value from the child iframe.");
        done();
    };
};

</script>
<iframe id="iframe" src="resources/multiple-suborigins.php"></iframe>
